# pocketllm-backend/app/api/v1/endpoints/users.py
"""User profile endpoints."""

from __future__ import annotations

from uuid import UUID

from fastapi import APIRouter, Depends, File, HTTPException, UploadFile, status

from app.api.deps import (
    get_current_request_user,
    get_database_dependency,
    get_settings_dependency,
)
from app.schemas.auth import TokenPayload
from app.schemas.users import (
    CancelDeletionResponse,
    DeleteAccountResponse,
    OnboardingSurvey,
    UserProfile,
    UserProfileUpdate,
)
from app.services.assets import UserAssetService
from app.services.users import UsersService

router = APIRouter(prefix="/users", tags=["users"])


@router.get("/profile", response_model=UserProfile, summary="Get user profile")
async def get_profile(
    user: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> UserProfile:
    service = UsersService(database=database)
    return await service.get_profile(user.sub)


@router.put("/profile", response_model=UserProfile, summary="Update user profile")
async def update_profile(
    payload: UserProfileUpdate,
    user: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> UserProfile:
    service = UsersService(database=database)
    return await service.update_profile(user.sub, payload)


@router.delete("/profile", response_model=DeleteAccountResponse, summary="Delete user account")
async def delete_profile(
    user: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> DeleteAccountResponse:
    service = UsersService(database=database)
    return await service.schedule_deletion(user.sub)


@router.post(
    "/profile/deletion/cancel",
    response_model=CancelDeletionResponse,
    summary="Cancel scheduled account deletion",
)
async def cancel_account_deletion(
    user: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> CancelDeletionResponse:
    service = UsersService(database=database)
    return await service.cancel_deletion(user.sub)


@router.post("/profile/onboarding", response_model=UserProfile, summary="Complete onboarding survey")
async def complete_onboarding(
    payload: OnboardingSurvey,
    user: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> UserProfile:
    service = UsersService(database=database)
    return await service.complete_onboarding(user.sub, payload)


@router.post(
    "/profile/avatar",
    response_model=UserProfile,
    summary="Upload or update the signed-in user's avatar image",
)
async def upload_profile_avatar(
    file: UploadFile = File(...),
    user: TokenPayload = Depends(get_current_request_user),
    settings=Depends(get_settings_dependency),
    database=Depends(get_database_dependency),
) -> UserProfile:
    service = UserAssetService(settings=settings, database=database)
    return await service.upload_avatar(user.sub, file)


@router.get("/{user_id}", response_model=UserProfile, summary="Get user profile by ID")
async def get_user_by_id(
    user_id: UUID,
    requester: TokenPayload = Depends(get_current_request_user),
    database=Depends(get_database_dependency),
) -> UserProfile:
    if requester.sub != user_id:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Cannot access another user's profile")
    service = UsersService(database=database)
    return await service.get_profile_by_id(user_id)